+++ src/lib/Containers/Conversation.tsx
@@ -1,19 +1,11 @@
+import { MarkdownString } from "danger/distribution/dsl/Aliases"
 import * as React from "react"
-import {
-  commitMutation,
-  createRefetchContainer,
-  Environment,
-  graphql,
-  MutationConfig,
-  RecordSourceSelectorProxy,
-  RelayRefetchProp,
-} from "react-relay"
+import { createFragmentContainer, graphql, RelayPaginationProp } from "react-relay"
 import { ConnectionHandler } from "relay-runtime"
 
 import { MetadataText, SmallHeadline } from "../Components/Inbox/Typography"
 
-import { FlatList, ImageURISource, NetInfo, View, ViewProperties } from "react-native"
-import ReversedFlatList from "react-native-reversed-flat-list"
+import { ActivityIndicator, FlatList, ImageURISource, NetInfo, View, ViewProperties } from "react-native"
 
 import styled from "styled-components/native"
 import colors from "../../data/colors"
@@ -21,9 +13,8 @@ import fonts from "../../data/fonts"
 import ConnectivityBanner from "../Components/ConnectivityBanner"
 
 import Composer from "../Components/Inbox/Conversations/Composer"
-import Message from "../Components/Inbox/Conversations/Message"
-import ArtworkPreview from "../Components/Inbox/Conversations/Preview/ArtworkPreview"
-import ShowPreview from "../Components/Inbox/Conversations/Preview/ShowPreview"
+import Messages from "../Components/Inbox/Conversations/Messages"
+import { sendConversationMessage } from "../Components/Inbox/Conversations/SendConversationMessage"
 
 import { markLastMessageRead } from "../Components/Inbox/Conversations/MarkReadMessage"
 
@@ -70,16 +61,19 @@ const MessagesList = styled(FlatList)`
   margin-top: 10;
 `
 
-const PAGE_SIZE = 100
+const LoadingIndicator = styled(ActivityIndicator)`
+  margin-top: 20;
+`
 
 interface Props extends RelayProps {
-  relay?: RelayRefetchProp
+  relay?: RelayPaginationProp
 }
 
 interface State {
   sendingMessage: boolean
   isConnected: boolean
   markedMessageAsRead: boolean
+  fetchingData: boolean
 }
 
 export class Conversation extends React.Component<Props, State> {
@@ -91,41 +85,22 @@ export class Conversation extends React.Component<Props, State> {
       sendingMessage: false,
       isConnected: true,
       markedMessageAsRead: false,
+      fetchingData: false,
     }
     this.handleConnectivityChange = this.handleConnectivityChange.bind(this)
   }
 
-  renderMessage({ item }) {
-    const conversationItem = this.props.me.conversation.items[0].item
-    const conversation = this.props.me.conversation
-    const partnerName = conversation.to.name
-    const senderName = item.is_from_user ? conversation.from.name : partnerName
-    const initials = item.is_from_user ? conversation.from.initials : conversation.to.initials
-    return (
-      <Message
-        message={item}
-        firstMessage={item.first_message}
-        initialText={conversation.initial_message}
-        senderName={senderName}
-        initials={initials}
-        artworkPreview={
-          item.first_message &&
-          conversationItem.__typename === "Artwork" &&
-          <ArtworkPreview
-            artwork={conversationItem}
-            onSelected={() => ARSwitchBoard.presentNavigationViewController(this, conversationItem.href)}
-          />
-        }
-        showPreview={
-          item.first_message &&
-          conversationItem.__typename === "Show" &&
-          <ShowPreview
-            show={conversationItem}
-            onSelected={() => ARSwitchBoard.presentNavigationViewController(this, conversationItem.href)}
-          />
-        }
-      />
-    )
+  componentDidMount() {
+    NetInfo.isConnected.addEventListener("connectionChange", this.handleConnectivityChange)
+    this.maybeMarkLastMessageAsRead()
+  }
+
+  componentWillUnmount() {
+    NetInfo.isConnected.removeEventListener("connectionChange", this.handleConnectivityChange)
+  }
+
+  handleConnectivityChange(isConnected) {
+    this.setState({ isConnected })
   }
 
   maybeMarkLastMessageAsRead() {
@@ -146,26 +121,9 @@ export class Conversation extends React.Component<Props, State> {
     }
   }
 
-  componentDidMount() {
-    NetInfo.isConnected.addEventListener("connectionChange", this.handleConnectivityChange)
-    this.maybeMarkLastMessageAsRead()
-  }
-
-  componentWillUnmount() {
-    NetInfo.isConnected.removeEventListener("connectionChange", this.handleConnectivityChange)
-  }
-
-  handleConnectivityChange(isConnected) {
-    this.setState({ isConnected })
-  }
-
   render() {
     const conversation = this.props.me.conversation
     const partnerName = conversation.to.name
-    const messages = conversation.messages.edges.map((edge, index) => {
-      return { first_message: index === 0, key: index, ...edge.node }
-    })
-    const lastMessage = messages[messages.length - 1]
 
     return (
       <Composer
@@ -195,13 +153,14 @@ export class Conversation extends React.Component<Props, State> {
               </SmallHeadline>
               <PlaceholderView />
             </HeaderTextContainer>
+            <LoadingIndicator animating={this.state.fetchingData} hidesWhenStopped />
           </Header>
           {!this.state.isConnected && <ConnectivityBanner />}
-          <ReversedFlatList
-            data={messages}
-            renderItem={this.renderMessage.bind(this)}
-            length={messages.length}
-            ItemSeparatorComponent={DottedBorder}
+          <Messages
+            conversation={conversation}
+            onDataFetching={loading => {
+              this.setState({ fetchingData: loading })
+            }}
           />
         </Container>
       </Composer>
@@ -209,157 +168,23 @@ export class Conversation extends React.Component<Props, State> {
   }
 }
 
-function sendConversationMessage(
-  environment: Environment,
-  conversation: RelayProps["me"]["conversation"],
-  text: string,
-  onCompleted: MutationConfig["onCompleted"],
-  onError: MutationConfig["onError"]
-) {
-  const lastMessage = conversation.messages.edges[conversation.messages.edges.length - 1].node
-
-  const storeUpdater = (store: RecordSourceSelectorProxy) => {
-    const mutationPayload = store.getRootField("sendConversationMessage")
-    const newMessageEdge = mutationPayload.getLinkedRecord("messageEdge")
-    const connection = ConnectionHandler.getConnection(store.get(conversation.__id), "Conversation_messages")
-    ConnectionHandler.insertEdgeAfter(connection, newMessageEdge)
-  }
-
-  return commitMutation(environment, {
-    onCompleted,
-    onError,
-
-    optimisticUpdater: storeUpdater,
-    updater: storeUpdater,
-
-    // TODO: See if we can extract the field selections into a fragment and share it with the normal pagination fragment.
-    //      Also looks like we can get rid of the `body` selection.
-    mutation: graphql`
-      mutation ConversationSendMessageMutation($input: SendConversationMessageMutationInput!) {
-        sendConversationMessage(input: $input) {
-          messageEdge {
-            node {
-              impulse_id
-              is_from_user
-              body
-              __id
-              ...Message_message
-            }
-          }
-        }
-      }
-    `,
-
-    variables: {
-      input: {
-        id: conversation.id,
-        from: conversation.from.email,
-        body_text: text,
-        // Reply to the last message
-        reply_to_message_id: lastMessage.impulse_id,
-      },
-    },
-
-    // TODO: Figure out which of these keys is *actually* required for Relay Modern and update the typings to reflect that.
-    //      And if it’s really true that this config isn’t enough to update the connection and we really need the updater
-    //      functions.
-    configs: [
-      {
-        type: "RANGE_ADD",
-        parentName: "conversation",
-        parentID: "__id",
-        connectionName: "messages",
-        edgeName: "messageEdge",
-        rangeBehaviors: {
-          "": "append",
-        },
-        connectionInfo: [
-          {
-            key: "Conversation_messages",
-            rangeBehavior: "append",
-          },
-        ],
-      },
-    ],
-
-    optimisticResponse: {
-      sendConversationMessage: {
-        messageEdge: {
-          node: {
-            body: text,
-            from: {
-              email: conversation.from.email,
-              name: null,
-            },
-            is_from_user: true,
-            created_at: null, // Intentionally left blank so Message can recognize this as an optimistic response.
-            attachments: [],
-          },
-        },
-      },
-    },
-  })
-}
-
-// TODO: Make this a pagination container instead of fetching 100 messages
-export default createRefetchContainer(
-  Conversation,
-  graphql`
+export default createFragmentContainer(Conversation, {
+  me: graphql`
     fragment Conversation_me on Me {
       conversation(id: $conversationID) {
-        __id
-        id
-        from {
-          name
-          email
-          initials
-        }
         to {
           name
           initials
         }
+        ...Messages_conversation
         initial_message
         is_last_message_to_user
         last_message_open
         last_message_delivery_id
-        messages(first: 200) @connection(key: "Conversation_messages") {
-          pageInfo {
-            hasNextPage
-          }
-          edges {
-            node {
-              impulse_id
-              is_from_user
-              body
-              ...Message_message
-            }
-          }
-        }
-        items {
-          item {
-            ... on Artwork {
-              __typename
-              href
-              ...ArtworkPreview_artwork
-            }
-            ... on Show {
-              __typename
-              href
-              ...ShowPreview_show
-            }
-          }
-        }
       }
     }
   `,
-  graphql`
-    query ConversationRefetchQuery($conversationID: String!) {
-      me {
-        ...Conversation_me
-      }
-    }
-  `
-)
+})
 
 export interface RelayProps {
   me: {
@@ -375,9 +200,6 @@ export interface RelayProps {
         name: string
         initials: string
       }
-      items: Array<{
-        item: any
-      }>
       initial_message: string
       is_last_message_to_user: boolean
       last_message_open: string | null
@@ -393,6 +215,9 @@ export interface RelayProps {
           } | null
         }>
       }
+      items: Array<{
+        item: any
+      }>
     }
   }
 }